refactor(FilterTypeFileTypes): extract search filter to useContentTypeSearchFilter composable
