+++ src/components/Filter/FilterType/FilterTypeFileTypes.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { computed, ref, toRef, useTemplateRef } from 'vue'
-import { deburr, isEqual } from 'lodash'
+import { isEqual } from 'lodash'
 
 import ButtonToggleContentTypesView from '@/components/Button/ButtonToggleContentTypesView.vue'
 import ContentTypesAll from '@/components/ContentTypes/ContentTypesCategories/ContentTypesAll.vue'
@@ -15,6 +15,7 @@ import settings from '@/utils/settings'
 import { getDocumentTypeLabel } from '@/utils/utils'
 import { useContentTypeCategories } from '@/composables/useContentTypeCategories'
 import { useContentTypeCategoryLabel } from '@/composables/useContentTypeCategoryLabel'
+import { useContentTypeSearchFilter } from '@/composables/useContentTypeSearchFilter'
 import { useSearchFilter } from '@/composables/useSearchFilter'
 import { useSearchStore } from '@/store/modules'
 import { CONTENT_TYPE_CATEGORY_FILTER_NAME } from '@/store/filters/FilterContentTypeCategory'
@@ -27,7 +28,6 @@ const props = defineProps({
 })
 
 const grouped = ref(true)
-const query = ref('')
 
 const filterTypeRef = useTemplateRef('filterTypeRef')
 const entries = computed(() => filterTypeRef.value?.entries ?? [])
@@ -45,66 +45,17 @@ const {
 const searchStore = useSearchStore.inject()
 const categoryLabelFor = useContentTypeCategoryLabel()
 
-const normalizeForSearch = value => deburr(String(value ?? '')).toLowerCase()
-const normalizedQuery = computed(() => normalizeForSearch(query.value))
-const hasQuery = computed(() => normalizedQuery.value !== '')
-
-// Precompute deburred haystacks once per (entries/categories/locale) change so
-// the per-keystroke filter doesn't redo deburr+lowercase across every label
-// and MIME key on every render.
-const categoryHaystacks = computed(() => {
-  const map = new Map()
-  Object.keys(categories.value).forEach((category) => {
-    map.set(category, normalizeForSearch(categoryLabelFor(category)))
-  })
-  return map
+const {
+  query,
+  visibleTypesFor,
+  visibleEntries,
+  filteredCategoryPairs
+} = useContentTypeSearchFilter({
+  contentTypes,
+  categories,
+  categoryLabelFor,
+  isContentTypeRetained: contentType => isEntryRetainedDuringSearch(contentType)
 })
-
-const contentTypeHaystacks = computed(() => {
-  const map = new Map()
-  contentTypes.value.forEach((contentType) => {
-    const key = normalizeForSearch(contentType)
-    const label = normalizeForSearch(getDocumentTypeLabel(contentType))
-    map.set(contentType, `${key}\n${label}`)
-  })
-  return map
-})
-
-const matchesContentType = (contentType) => {
-  if (!hasQuery.value) {
-    return true
-  }
-  return contentTypeHaystacks.value.get(contentType)?.includes(normalizedQuery.value) ?? false
-}
-
-const matchesCategory = (category) => {
-  if (!hasQuery.value) {
-    return true
-  }
-  return categoryHaystacks.value.get(category)?.includes(normalizedQuery.value) ?? false
-}
-
-// Keep selected and category-implied types visible while typing so the user
-// doesn't appear to "lose" a selection.
-const isContentTypeVisible = contentType =>
-  matchesContentType(contentType) || isEntryRetainedDuringSearch(contentType)
-
-// A category-label hit short-circuits per-type filtering: every type the API
-// returned for that category stays visible, even those that don't individually
-// match the query.
-const visibleTypesFor = (category, types) => {
-  if (!hasQuery.value || matchesCategory(category)) {
-    return types
-  }
-  return types.filter(isContentTypeVisible)
-}
-
-const visibleEntries = (slotEntries) => {
-  if (!hasQuery.value) {
-    return slotEntries
-  }
-  return slotEntries.filter(entry => isContentTypeVisible(entry.item.key))
-}
 
 const sort = computed(() => searchStore.sortFilters[props.filter.name] ?? {
   sortBy: settings.filter.sortBy,
@@ -257,21 +208,6 @@ const toggleEntry = (contentType, checked) => {
 
 const categoryCount = types =>
   types.reduce((sum, contentType) => sum + entryCount(contentType), 0)
-
-// Empty-query fast path returns the original pairs ref so the grouped view
-// re-renders identically when no filter is active.
-const filteredCategoryPairs = computed(() => {
-  const pairs = Object.entries(categories.value)
-  if (!hasQuery.value) {
-    return pairs
-  }
-  return pairs.filter(([category, types]) => {
-    if (matchesCategory(category)) {
-      return true
-    }
-    return types.some(isContentTypeVisible)
-  })
-})
 
 const sortedCategoryEntries = computed(() => {
   const pairs = filteredCategoryPairs.value
